% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtess.R
\name{as.polygons}
\alias{as.polygons}
\title{Convert the Dirichlet tessellation to polygons and clip if so desired.}
\usage{
as.polygons(x, bboxclip = TRUE, proj = sp::CRS("EPSG:4326"), ...)
}
\arguments{
\item{x}{Dirichlet tessellation object}

\item{bboxclip}{Logical. Clip to the bounding box of the coordinates used to
create the tessellations.}

\item{proj}{CRS Projection of coordinates default is "EPSG:4326" (lon/lat wgs84)}
}
\description{
Convert the Dirichlet tessellation to polygons and clip if so desired.
}
\examples{
library(sp)
coords <- matrix(runif(2000),ncol=2)
tess <- dirtess(coords)
tess.polys <- as.polygons(tess)
plot(tess.polys$polygons, col=hcl.colors(100))
}
