% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_utils.R
\name{as.spatstat}
\alias{as.spatstat}
\title{Function to covert a spatstat quadrature and spatstat data (covariates)}
\usage{
as.spatstat(object, ...)
}
\arguments{
\item{object}{A ppmData object}

\item{\\dots}{Ignored}
}
\description{
This function will convert the ppmData to a spatstat object that
can be used in the spatstat library. This currenly only works for unmarked
point process
}
\examples{
\dontrun{
library(ppmData)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
preds <- rast(lst)
window <- preds[[1]]
presences <- subset(snails,SpeciesID \%in\% "Tasmaphena sinclairi")
object <- ppmData(npoints = 10000, presences=presences, window = window, covariates = preds)
ss.object <- as.spatstat(object)

## now you should be able to fit a ppm from spatstat
ss.form <- ss.object$Q~1+poly(min_temp_coldest_month,2)+poly(annual_mean_precip,2)
ppmfit <- ppm(ss.form,data=ss.object$covariates,Poisson())
pred <- predict(ppmfit, covariates = ss.object$covariates,type="trend")
}
}
