% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Non-parametric bootstrap to assess uncertainty in ppm parameters}
\usage{
bootstrap(object, nboot = 10, quiet = TRUE, mc.cores = 1, ...)
}
\arguments{
\item{object}{A ppm model object}

\item{nboot}{Default is 10; The number of bootstraps to run.}

\item{quiet}{Default is TRUE; Run the bootstrap with out any trace.}

\item{mc.cores}{Default is 1; the number of cores if run in parallel}

\item{\dots}{Ignored}
}
\description{
This function uses a non-parametric bootstrap to estimate
uncertainty in a ppm. Conditional on X_j = {X_i,...,X_n}, let N*
have a Poisson distribution with Lambda_j. Lambda_j can be estimated as the
integral for each species intensities (lambda_ij), conditional on the
archetypes. For each species we can then draw \eqn{X*_{1},...,X*_{N}} by
sampling randomly with replacement N* times from X. We refit the ppm with
this new realisation of the ppm data. Estimated coefficients are returned
from each of B bootstrap model fits. This bootstrap object can then be used
in the to predict \link[stats]{predict} ppm and generate confidence intervals
function. If the model is very large (many po observations + quadrature sites)
vthis will be slow, and will take nboot times the time it takes to fit a
single ppm with the same data. So be warned for large datasets. We implement
method two from Cowling et al., (1996).
}
\references{
Cowling, A., Hall, P. and Phillips, M.J., 1996. Bootstrap
confidence regions for the intensity of a Poisson point process. Journal of
the American Statistical Association, 91(436), pp.1516-1524.

Woolley, S.N.C., Dunstan, P.K., Warton, D., and Foster, S.,
InPrep. Multiple species modelling of presence-only data-sets with
mixture-of-regressions models.
}
