% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{cpp_to_df}
\alias{cpp_to_df}
\title{sometimes vectors come back as col-wise from c++ and need to be make row-wise for R.}
\usage{
cpp_to_df(x)
}
\description{
sometimes vectors come back as col-wise from c++ and need to be make row-wise for R.
}
