% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltri.R
\name{deltri}
\alias{deltri}
\title{deltri}
\usage{
deltri(coords)
}
\arguments{
\item{coords}{The coordinates of the points to triangulate.}
}
\description{
Computes the Delaunay triangulation for large data sets. The
main use of this function is to quickly calculate the Delaunay triangulation
so we can calculate the duel-graph and Dirichlet tessellation to use in point
process modelling, where the area of each Dirichlet tessellation is used as a
areal weight in a statistical model.
}
\examples{
coordin <- matrix(runif(2000),ncol=2)
tri <- deltri(coordin)
}
