% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtess.R
\name{dirTess}
\alias{dirTess}
\title{dirtess}
\usage{
dirTess(coords, bbox = NULL)
}
\arguments{
\item{coords}{The coordinates of the points to triangulate.}

\item{bbox}{Optional - doesn't actually do anything anymore :)}
}
\description{
Computes the Dirichlet tessellation using the dual-graph of the
Delaunay triangulation. The Delaunay triangulation is calculated internally
for this function using a sweep algorithm.
}
\examples{
coords <- matrix(runif(2000),ncol=2)
tess <- dirTess(coords)
}
