% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{eip}
\alias{eip}
\title{eip (estimate inclusion probabilties) creates a probability surface based on
 the location of known survey sites}
\usage{
eip(known.sites, study.area = NULL, proj = NULL, params)
}
\arguments{
\item{known.sites}{a matrix, dataframe or SpatialPoints* object giving the
coordinates of the points to use in sampling (of size Nxdimension).
Note: SpatialPoints* will only be suitiable for \code{dimension=2}.}

\item{study.area}{an optional extent, SpatialPolygons* or Raster* object giving the
area over which to generate background points. If ignored, a rectangle
defining the extent of \code{known.sites} will be used instead.}

\item{probs}{a vector specifying the inclusion probability for each of the N sampling sites}

\item{pfr}{(projection for raster) a equal area projection appropriate for the region. This will be
used when setting up the ppp density layer.}

\item{sigma}{A numeric value or a function to estimate sigma see \code{\link[spatstat]{density.ppp}}
for details.}
}
\description{
Estimating the probabilty of presence from a series of spatial points.
The probability of *absence in an area of size A* according to the poisson distribution is
 \deqn{pr(y=0) = exp(-\lambda(u)*A)}
 The prob of *presence* is then
 \deqn{pr( y=1) = 1-pr(y=0)
                = 1-exp(-\lambda(u)*A)}
\eqn{\lambda(u)} = the intensity value at point \eqn{u}. This is estimated using \code{\link[spatstat]{density}}
}

