% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_grid.R
\name{gridWeights}
\alias{gridWeights}
\title{gridWeights}
\usage{
gridWeights(
  presences,
  quadrature,
  window,
  coord = c("X", "Y"),
  mark.id,
  unit = c("geo", "m", "km", "ha")
)
}
\arguments{
\item{presences}{The observed presence locations}

\item{quadrature}{The quadrature locations}

\item{window}{A SpatRaster from terra package which will represent the extent
and resolution of the point process model.}

\item{coord}{The names of the coordinates. Default is c("X","Y").}

\item{mark.id}{character Name of the species ID column}

\item{unit}{The scale of the area weights, default is kilometers squared "km"
but meters squared "m" or hectars "ha" can be used.}
}
\description{
Quick way to generate psuedo-random points using the terria
package. Weights are taken as the the area(window)/npoints. The default unit
is km^2, but other units can be used such as meters squared "m" or hectars
"ha". There is an internal check to see if the input data is in lon/lat and
if so, it returns the area of the cell in which the point lies.
}
\examples{
library(ppmData)
library(terra)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
window <- terra::rast(lst[1])
presences <- subset(snails,SpeciesID \%in\% "Tasmaphena sinclairi")[,1:2]
quadrature <- gridQuad(10000,window,control=list(quiet=FALSE))
res <- gridWeights(presences,quadrature,window)
plot(window)
points(res[res$presence==1,1:2],pch=16,cex=0.5,col='blue')
points(res[res$presence==0,1:2],pch='.')
}
\author{
Skipton Woolley
}
