% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtess.R
\name{plot.dirTess}
\alias{plot.dirTess}
\title{plot a Dirichlet tessellation}
\usage{
\method{plot}{dirTess}(x, ...)
}
\arguments{
\item{x}{dirTess object}

\item{\\dots}{Additional plotting arguments}
}
\description{
plot a Dirichlet tessellation
}
