% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirtess.R
\name{polygonise}
\alias{polygonise}
\title{Convert the Dirichlet tessellation to polygons and clip if so desired.}
\usage{
polygonise(x, clip = TRUE, polyclip = NULL, proj = sp::CRS("EPSG:4326"), ...)
}
\arguments{
\item{x}{Dirichlet tessellation object}

\item{clip}{Logical Clip to the bounding box or polgon if polyclip = TRUE.}

\item{polyclip}{polygon A polygon to clip the dirichlet tessellation - intersection between boundary polygons and polygon}

\item{proj}{CRS Projection of coordinates default is "EPSG:4326" (lon/lat wgs84)}

\item{\\dots}{Additional arguments for a polygonise function}
}
\description{
Convert the Dirichlet tessellation to polygons and clip if so desired.
}
\examples{
library(sp)
coords <- matrix(runif(2000),ncol=2)
tess <- dirTess(coords)
tess.polys <- polygonise(tess)
plot(tess.polys$polygons, col=hcl.colors(100))
}
