% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmdata.R
\name{ppmData.control}
\alias{ppmData.control}
\title{Controls for ppmData generation.}
\usage{
ppmData.control(
  quiet = FALSE,
  cores = 1,
  maxpoints = 250000,
  extractNArm = TRUE,
  extractBuffer = NULL,
  quasiSamps = 5000,
  quasiDims = 2,
  singlespeciesFormat = "long",
  multispeciesFormat = "wide",
  ...
)
}
\arguments{
\item{quiet}{Should any reporting be performed? Default is FALSE, for reporting.}

\item{cores}{The number of cores to use in fitting of species_mix models. These will be largely used to model the species-specific parameteres.}

\item{\dots}{Other control calls.}
}
\description{
Controls for ppmData generation.
}
