% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltri.R
\name{print.delTri}
\alias{print.delTri}
\title{Printing function for delTri}
\usage{
\method{print}{delTri}(x, ...)
}
\arguments{
\item{x}{A delTri object}

\item{\\dots}{Additional print calls.}
}
\description{
Printing function for delTri
}
