% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmdata.R
\name{print.ppmdata}
\alias{print.ppmdata}
\title{Print a summary of ppmdata object.}
\usage{
\method{print}{ppmdata}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{\\dots}{Ignored}
}
\description{
Print a summary of ppmdata object.
}
