% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_pseudo.R
\name{pseudoRandomQuad}
\alias{pseudoRandomQuad}
\title{psuedoRandomQuad}
\usage{
pseudoRandomQuad(npoints, window, coord = c("X", "Y"))
}
\arguments{
\item{npoints}{The number of background points to use. The default is 1000,
but it is generally recommended to use more when fitting a ppm.}

\item{window}{A SpatRaster from terra package which will represent the extent
and resolution of the point process model.}

\item{coord}{The names of the coordinates. Default is c("X","Y").}
}
\description{
A relatively quick way to generate psuedo-random quadrature set
for ppm models.
}
\examples{
library(ppmData)
library(terra)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
window <- terra::rast(lst[1])
res <- psuedoRandomQuad(window,10000)
plot(window)
points(res[,1:2],pch=".")
}
\author{
Skipton Woolley
}
