% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{rast_to_sf}
\alias{rast_to_sf}
\title{Convert a terra SpatRaster* object to an sf object
Convert a terra SpatRaster* object to an sf object.
This creates a binary mask for the raster NA vs data and then converts it to
POLYGONS. For use internally to turn a rast object in to a mask for polygon
clipping.}
\usage{
rast_to_sf(x, ...)
}
\arguments{
\item{x}{rast* to be converted into an object class \code{sf}}

\item{...}{further specifications, see \link{st_as_sf}}
}
\description{
Convert a terra SpatRaster* object to an sf object
Convert a terra SpatRaster* object to an sf object.
This creates a binary mask for the raster NA vs data and then converts it to
POLYGONS. For use internally to turn a rast object in to a mask for polygon
clipping.
}
