% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snails}
\alias{snails}
\title{Presence only observations of snails from the family Rhytididae located in Tasmania, Australia.}
\format{
A data frame with columns:
\describe{
 \item{X}{Longitude.}
 \item{Y}{Latitude.}
 \item{SpeciesID}{Species taxonomic name.}
}
}
\source{
Atlas of Living Australia. \url{https://www.ala.org/}
}
\usage{
snails
}
\description{
Presence only observations of snails from the family Rhytididae located in Tasmania, Australia.
}
\examples{
\dontrun{
 snails
}
}
\keyword{datasets}
