% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{terra2im}
\alias{terra2im}
\title{convert at terra raster to a spatstat image taken from maptools and converted to terra}
\usage{
terra2im(rast, factor.col.name = NULL)
}
\arguments{
\item{rast}{rast A terra raster to convert to a spatstat image}

\item{factor.col.name}{string Names of factors in the raster to convert to window}
}
\description{
convert at terra raster to a spatstat image taken from maptools and converted to terra
}
\examples{
library(ppmData)
library(terra)
path <- system.file("extdata", package = "ppmData")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
covariates <- rast(lst[1])
im.r <- terra2im(covariates)

plot(im.r)
}
