% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_background_points.R
\name{generate_background_points}
\alias{generate_background_points}
\title{create background points for study area or point data}
\usage{
generate_background_points(
  npoints = 10000,
  presences = NULL,
  window = NULL,
  covariates = NULL,
  resolution = 1,
  method = c("grid", "quasirandom", "random"),
  interpolation = "bilinear",
  coords = c("X", "Y")
)
}
\arguments{
\item{npoints}{number of back ground points to create - only applicable to quasirandom.}

\item{presences}{a matrix, dataframe or SpatialPoints* object giving the
coordinates of each species' presence in (should be a matrix of nsites * 3) with the three columns being c("X","Y","SpeciesID"), where X is longitude, Y is latitude and SpeciesID is a integer, character or factor which assoicates each point to a species.}

\item{window}{a Raster* object giving the area over which to generate background points. NA cells afe ignored.
If ignored, a rectangle defining the extent of \code{presences} will be used.}

\item{covariates}{an optional Raster* object containing covariates for
modelling the point process (best use a Raster* stack or Raster* brick)}

\item{resolution}{resolution setup grid for integration points (default is 1 deg).}

\item{method}{the type of method that should be used to generate background points. The options are:
'grid' generates a regular grid of background points. See Berman & Turner 1993 or Warton & Shepard 2010 for details.
'quasirandom' generates quasirandom background points. See Foster 2015 for details.
'random' generates a random set of background points - See Philips 2006 (ala MaxEnt) for details.}

\item{interpolation}{either 'simple' or 'bilinear' and this determines the interpolation method for selecting covariate data. 'simple' is nearest neighbour, 'bilinear' is bilinear interpolation.}

\item{coords}{is the name of site coordinates. The default is c('X','Y').}
}
\description{
create background points for study area or point data
}
