% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_background_points.R
\name{generate_background_points}
\alias{generate_background_points}
\title{create background points for study area or point data}
\usage{
generate_background_points(number_of_background_points = 2000,
  known_sites = NULL, study_area = NULL, model_covariates = NULL,
  resolution = 1, multispecies_points = NULL, bias_layer = NULL,
  method = c("grid", "quasirandom", "quasirandom_covariates", "multispecies",
  "bias"))
}
\arguments{
\item{number_of_background_points}{number of back ground points to create - only applicable to quasirandom.}

\item{known_sites}{a matrix, dataframe or SpatialPoints* object giving the
coordinates of the points to use in sampling (of size Nxdimension).
Note: SpatialPoints* will only be suitiable for \code{dimension=2}.}

\item{study_area}{a Raster* object giving the area over which to generate background points.
If ignored, a rectangle defining the extent of \code{known_sites} will be used.}

\item{model_covariates}{an optional Raster* object containing covariates for
modelling the point process (best use a Raster* stack or Raster* brick)}

\item{resolution}{resolution setup grid for integration points (default is 1 deg).}

\item{multispecies_points}{a community matrix of speciesXsites collection methods - still need to think about how to set this up.
probably need to setup an index to identify collection method.
eg. PO abs = 0, PO pres = 1, PA abs = 2, PA pres = 3, it would be easy to then assign any new collection methods with
PA-trawl-presence = 4, PA-trawl-abs = 5}

\item{method}{the type of method that should be used to generate background points.
'grid' generates a regular grid at a set resolution.
'quasirandom' generates quasirandom background points based on the coordinates.
'quasirandom_covariates' generates quasirandom background points across the multiple dimensions of covariate hyperdimensions.
'multispecies' more to follow soon.}
}
