% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ppmData}
\alias{plot.ppmData}
\title{Plot a ppmData object}
\usage{
\method{plot}{ppmData}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{\\dots}{Ignored}
}
\description{
Plot a ppmData object
}
