% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmdata.R
\name{ppmData}
\alias{ppmData}
\title{Create a Point Process dataset for spatial presence only modelling.}
\usage{
ppmData(
  npoints = 10000,
  presences = NULL,
  window = NULL,
  covariates = NULL,
  resolution = NULL,
  method = c("grid", "quasirandom", "random"),
  interpolation = "bilinear",
  coord = c("X", "Y"),
  control = list(maxpoints = 250000, extractNArm = TRUE, extractBuffer = NULL,
    quasiSamps = 5000, quasiDims = 2, multispeciesFormat = "wide")
)
}
\arguments{
\item{npoints}{number of background points to create.}

\item{presences}{a matrix, dataframe or SpatialPoints* object giving the coordinates of each species' presence in (should be a matrix of nsites * 3) with the three columns being c("X","Y","SpeciesID"), where X is longitude, Y is latitude and SpeciesID is a integer, character or factor which assoicates each point to a species. If presences are NULL then ppmdat will return the quadrature (background) points.}

\item{window}{a Raster* object giving the area over which to generate background points. NA cells are ignored and masked out of returned data.
If ignored, a rectangle defining the extent of \code{presences} will be used.}

\item{covariates}{an optional Raster object containing covariates for modelling the point process (best use a Raster stack or Raster brick).}

\item{resolution}{resolution setup grid for integration points (default is 1 deg) - but this need to be setup  with reference to original raster resolution.}

\item{method}{the type of method that should be used to generate background points. The options are:
'grid' generates a regular grid of background points. See Berman & Turner 1992 or Warton & Shepard 2010 for details.
'quasirandom' generates quasirandom background points. See Bratley & Fox 1998 or Foster etal 2015 for details.
'random' generates a random set of background points. See Philips 2006 (ala MaxEnt) for details.}

\item{interpolation}{either 'simple' or 'bilinear' and this determines the interpolation method for interpolating data across different cell resolutions.
'simple' is nearest neighbour, 'bilinear' is bilinear interpolation.}

\item{control}{is a list of options for generating quadrature scheme. Currently:
'maxpoints' = 250000 and sets a limit to number of integration points generated as background data.
'extractNArm' = TRUE and uses na.rm=TRUE for extracting raster covariate data.
'extractBuffer' = NULL and is the amount of buffer to provide each point on extract (radius from point).
'quasiSamps' = 5000 and is the default number of samples for halton random number generator.
'quasiDimss' = 2 and is the dimension to generate the random samples over > 3 you are looking at hyperdimensions.}

\item{SpeciesID}{is the name of site coordinates. The default is c('X','Y').}
}
\description{
Creates a point process data frame for modelling single species or multiple species (marked) presences. Generates a quadrature scheme based on Berman & Turner 1992; Warton & Shepard 2010. The function can generate a quadrature scheme for a regular grid, quasi-random or random points.
}
