% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmdata.R
\name{ppmData}
\alias{ppmData}
\title{Create a spatial point Process dataset for spatial modelling.}
\usage{
ppmData(
  npoints = 10000,
  presences = NULL,
  window = NULL,
  covariates = NULL,
  interpolation = "simple",
  coord = c("X", "Y"),
  mc.cores = parallel::detectCores() - 1,
  quasirandom.samples = NULL,
  quasirandom.dimensions = NULL
)
}
\arguments{
\item{npoints}{The number of background points to generate.}

\item{presences}{a matrix, dataframe or SpatialPoints* object giving the coordinates of each species' presence in (should be a matrix of nsites * 3) 
with the three columns being c("X","Y","SpeciesID"), where X is longitude, Y is latitude and SpeciesID is a integer, character or factor which assoicates each point to a species.
If presences parameter is NULL then ppmDat will return the quadrature (background) points.}

\item{window}{a Raster* object giving the area over which to generate background points. NA cells are ignored and masked out of returned data.
If ignored, a rectangle defining the extent of \code{presences} will be used.}

\item{covariates}{A Raster* object containing covariates for modelling the point process (best use a Raster stack or Raster brick).}

\item{interpolation}{either 'simple' or 'bilinear' and this determines the interpolation method for interpolating data across different cell resolutions.
'simple' is nearest neighbour, 'bilinear' is bilinear interpolation.}

\item{coord}{is the name of site coordinates. The default is c('X','Y').}

\item{mc.cores}{The number of cores to use in the processing to quasirandom points and weighting scheme.}

\item{quasirandom.samples}{This set the total number of samples to consider in the BAS step (rejection sampling). 
The default is 10000, which means that 10000 halton numbers are drawn and then thinned according to the inclusion probabilities.
You will need to increase the number of samples if sampling across > 2 dimensions or selecting a large number of background points. 
The more quasirandomSample selected the slower the background point generation will be.}

\item{quasirandom.dimensions}{The the number of dimensions that the samples are located in. Equal to 2 for areal sampling. 
Care should be taken with large dimensions as :1) the number of potential sampling sites needed for effective coverage starts to explode (curse of dimensionality);
and 2) the well-spaced behaviour of the Halton sequence starts to deteriorate.}
}
\description{
Creates a point process data object for modelling Point Process presence-only datasets. 
Generates a quadrature scheme based on Berman & Turner 1992; Warton & Shepard 2010. 
The function can generate a quadrature scheme for a regular grid, quasi-random or random points.
}
\examples{
library(qrbp)
path <- system.file("extdata", package = "qrbp")
lst <- list.files(path=path,pattern='*.tif',full.names = TRUE)
preds <- raster::stack(lst)
presences <- subset(snails,SpeciesID \%in\% "Tasmaphena sinclairi") 
bkgrid <- ppmData(npoints = 1000, presences=presences, window = preds[[1]], covariates = preds)
}
