% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ppmData}
\alias{print.ppmData}
\title{Print a summary of ppmData object.}
\usage{
\method{print}{ppmData}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{\\dots}{Ignored}
}
\description{
Print a summary of ppmData object.
}
